requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
autoenv==1.0.0
base==1.0.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coverage==5.1
dateparser==0.7.4
DateTime==4.3
Flask==1.1.2
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
pprintpp==0.4.0
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.9.0
pytest-mock==3.1.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
pytz-convert==0.2.9
regex==2020.5.14
requests==2.23.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
tzlocal==2.1
urllib3==1.25.9
wcwidth==0.2.3
Werkzeug==1.0.1
zope.interface==5.1.0
